refactor(database): отделить cli от функций
