+++ src/database/functions/clearDatabase.ts
@@ -1,6 +1,6 @@
 import { DataSource } from 'typeorm';
-import { dataSourceOptions } from '../typeorm.config';
-import { DEVELOPMENT } from '../common/constants';
+import { dataSourceOptions } from '../../typeorm.config';
+import { DEVELOPMENT } from '../../common/constants';
 import { Logger } from '@nestjs/common';
 
 export default async function clearDatabase(
